print: Create "print" module for printing directly from code
Create a module providing a print.printer model to represent printers
connected to the Odoo server.  This allows for backend code to send
reports directly to a printer, rather than requiring the user to
download a PDF to be printed manually.

The primary use case is for printing reports: this can be triggered in
the Python code via e.g.

  Printer = self.env['print.printer']
  Picking = self.env['stock.picking']
  pick = Picking.search([('name', '=', 'PICK00009')])
  Printer.spool_report(pick.ids, 'stock.report_deliveryslip')

or equivalently using XMLRPC via e.g.

  ids = models.execute_kw(dbname, uid, password,
                          'stock.picking', 'search',
                          [[['name', '=', 'PICK00009']]])
  models.execute_kw(dbname, uid, password,
                    'print.printer', 'spool_report',
                    [[], ids, 'stock.report_deliveryslip'])

Note that spool_report() may be called on a specific printer or
directly on the model.  If called directly on the model, the user's
default printer (or system default printer) will be used.

Another common use case is to identify a printer by barcode and
associate it as the current user's default printer:

  ids = models.execute_kw(dbname, uid, password, 'print.printer',
                          'search', [[['barcode', '=', 'PRNLAB02']]])
  models.execute_kw(dbname, uid, password, 'print.printer',
                    'set_user_default', [ids])

Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>