[FIX] print: Allow reinstallation after changing default printer
There is a SQL constraint ensuring that only one printer can ever be
marked as the system default.  If a user-created printer is marked as
the system default and the print module is reinstalled, then a unique
constraint violation will occur.

Fix by using a YAML file to set the "is_default" flag on
print.default_printer only if no other printer is already configured
as default.

Modified-by: Michael Brown <mbrown@fensystems.co.uk>
Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>