[IMP] print: Raise UserError for a non-existent report name
When an attempt is made to print a non-existent report, the
ir.actions.report._get_report_from_name() method will return an empty
recordset.  The subsequent call to report.render_qweb_pdf() will
return a backtrace that is spectacularly unhelpful in tracking down
the root cause of the problem.

Fix by explicitly raising UserError if the user attempts to print a
non-existent report.

Reported-by: Joe Burgess <joe.burgess@unipart.io>
Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>