[TEST] print: Add assertCpclReport() to verify expected CPCL/XML content
Add the assertCpclReport() method to allow the content of a generated
CPCL/XML report to be verified against a reference test file.  The
comparison uses a whitespace-stripping parser and the XML
canonicalization mechanism to perform a semantic comparison, and
pretty-prints the result to produce a meaningful and easy to read diff
in the event of a comparison failure.

Refernce test files are assumed to be in the <modulename>/tests/files
subdirectory, using code borrowed from the "edi" module.

Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>