[IMP] print: Add concept of a printer group
Add the standard Odoo parent/child model to print.printer, allowing
for the concept of a group of printers.  Define both per-user and
system-wide default printers on a per-group basis.

This allows callers to print to a printer group: the actual printer
selected will be based upon the user (or system) default printer
within that group.  This can be used to model a situation in which a
user requires multiple concepts of "default printer": for example, a
warehouse worker may need to have standard A4 PDF delivery slips
printed on an office laser printer but to have 4x4" labels printed on
a mobile Zebra direct thermal printer.

Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>