[IMP] print: Allow automatic selection of report type for a printer
Define a report type (e.g. "qweb-pdf" or "qweb-cpcl") for a printer
and allow spool_report() to be called with a recordset of multiple
reports, with the appropriate report being selected automatically for
each printer based on matching the report types.

This allows CPCL/XML test pages to be printed via the standard "Print
Test Page" action.

Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>