[TEST] print: Prevent ir.actions.report from releasing the test savepoint
When tests are run via the "-f" command-line option, the internal
"test_enable" flag is not set.  Via a remarkably convoluted chain of
events, this causes ir.actions.report.render_qweb_pdf() to set the
flag "commit_assetsbundle" within the context, which in turn causes
ir_qweb to explicitly commit the database cursor after generating the
assets bundle.  This breaks a fairly fundamental assumption of the
underlying SavepointCase, leading to an exciting failure when
SavepointCase attempts to roll back to the no longer existent
savepoint.

Fix by forcibly setting the "test_enable" flag for the duration of
printer test cases.

Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>