[IMP] print: Allow one user default printer per report type
User can now have a default printer for each report type, e.g. a default printer for PDF and another for CPCL reports.

Computed printer_id field removed on res_users and replaced with _get_printer function, which will look for the default printer for the specified report type, if set.

Updated spool_report and printers to work with the above changes, with explicit checks for a printer that matches the report type of the document that is being printed.

Story/10476

Signed-off-by: Peter Clark <peter.clark@unipart.io>