[ADD] print: Added unit test for one user default printer group per report type
Test added to check that a user can have one default printer per report type, and that changing the default of one type doesn't affect the user default printer settings of other types.

Added helper methods to create printers, check if a specific printer is the default for a specific user (as computed fields are unreliable in unit tests) and to assert that a printer is/isn't a default as expected.

Story/10476

Signed-off-by: Peter Clark <peter.clark@unipart.io>